requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
twitchAPI==2.5.4
typing_extensions==4.2.0
urllib3==1.26.9
websockets==10.3
yarl==1.7.2
